Add dummy test, refactor.
